setup: recognise extensions.objectFormat
This extension selects which hashing algorithm from vtable should be
used for reading and writing objects in the object store.  At the moment
it supports only single value (sha-1).

In case value of objectFormat is an unknown hashing algorithm, Git
command will fail with the following message:

  fatal: unknown repository extensions found:
	  objectformat = <value>

To indicate, that this specific objectFormat value is not recognised.

The objectFormat extension is not allowed in repository marked as
version 0 to prevent any possibility of accidentally writing a NewHash
object in the sha-1 object store. This extension behaviour is different
than preciousObjects extension (which is allowed in repo version 0).

Add tests and documentation note about new extension.

Signed-off-by: Patryk Obara <patryk.obara@gmail.com>